+++ packages/liveblocks-core/src/index.ts
@@ -328,6 +328,7 @@ export type {
   AiAssistantMessage,
   AiChat,
   AiChatMessage,
+  AiChatsQuery,
   AiKnowledgeSource,
   AiReasoningPart,
   AiTextPart,